Fix indexes for JsonDocument without subcolumns
