+++ ydb/core/kqp/ut/olap/ya.make
@@ -32,6 +32,7 @@ SRCS(
     sys_view_ut.cpp
     tiering_ut.cpp
     write_ut.cpp
+    kqp_olap_json_ut.cpp
 )
 
 PEERDIR(
